# swagger_server/models/baker_review.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from database.dbmodels.baker_review_dbmodel import BakerReviewDbModel
from swagger_server.models.base_model_ import Model
from swagger_server import util


class BakerReview(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, review_id: str=None, baker_id: str=None, author_id: str=None, rating: int=None, description: str=None):  # noqa: E501
        """BakerReview - a model defined in Swagger

        :param review_id: The review_id of this BakerReview.  # noqa: E501
        :type review_id: str
        :param baker_id: The baker_id of this BakerReview.  # noqa: E501
        :type baker_id: str
        :param author_id: The author_id of this BakerReview.  # noqa: E501
        :type author_id: str
        :param rating: The rating of this BakerReview.  # noqa: E501
        :type rating: int
        :param description: The description of this BakerReview.  # noqa: E501
        :type description: str
        """
        self.swagger_types = {
            'review_id': str,
            'baker_id': str,
            'author_id': str,
            'rating': int,
            'description': str
        }

        self.attribute_map = {
            'review_id': 'review_id',
            'baker_id': 'baker_id',
            'author_id': 'author_id',
            'rating': 'rating',
            'description': 'description'
        }
        self._review_id = review_id
        self._baker_id = baker_id
        self._author_id = author_id
        self._rating = rating
        self._description = description

    @classmethod
    def from_dict(cls, dikt) -> 'BakerReview':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The BakerReview of this BakerReview.  # noqa: E501
        :rtype: BakerReview
        """
        return util.deserialize_model(dikt, cls)

    @property
    def review_id(self) -> str:
        """Gets the review_id of this BakerReview.


        :return: The review_id of this BakerReview.
        :rtype: str
        """
        return self._review_id

    @review_id.setter
    def review_id(self, review_id: str):
        """Sets the review_id of this BakerReview.


        :param review_id: The review_id of this BakerReview.
        :type review_id: str
        """

        self._review_id = review_id

    @property
    def baker_id(self) -> str:
        """Gets the baker_id of this BakerReview.


        :return: The baker_id of this BakerReview.
        :rtype: str
        """
        return self._baker_id

    @baker_id.setter
    def baker_id(self, baker_id: str):
        """Sets the baker_id of this BakerReview.


        :param baker_id: The baker_id of this BakerReview.
        :type baker_id: str
        """

        self._baker_id = baker_id

    @property
    def author_id(self) -> str:
        """Gets the author_id of this BakerReview.


        :return: The author_id of this BakerReview.
        :rtype: str
        """
        return self._author_id

    @author_id.setter
    def author_id(self, author_id: str):
        """Sets the author_id of this BakerReview.


        :param author_id: The author_id of this BakerReview.
        :type author_id: str
        """
        if author_id is None:
            raise ValueError("Invalid value for `author_id`, must not be `None`")  # noqa: E501

        self._author_id = author_id

    @property
    def rating(self) -> int:
        """Gets the rating of this BakerReview.


        :return: The rating of this BakerReview.
        :rtype: int
        """
        return self._rating

    @rating.setter
    def rating(self, rating: int):
        """Sets the rating of this BakerReview.


        :param rating: The rating of this BakerReview.
        :type rating: int
        """
        allowed_values = ["1", "2", "3", "4", "5", "6", "7", "8", "9", "10"]  # noqa: E501
        if rating not in allowed_values:
            raise ValueError(
                "Invalid value for `rating` ({0}), must be one of {1}"
                .format(rating, allowed_values)
            )

        self._rating = rating

    @property
    def description(self) -> str:
        """Gets the description of this BakerReview.


        :return: The description of this BakerReview.
        :rtype: str
        """
        return self._description

    @description.setter
    def description(self, description: str):
        """Sets the description of this BakerReview.


        :param description: The description of this BakerReview.
        :type description: str
        """
        if description is None:
            raise ValueError("Invalid value for `description`, must not be `None`")  # noqa: E501

        self._description = description

    def toBakerReviewDbModel(self):
        baker_review_dbmodel = BakerReviewDbModel()
        baker_review_dbmodel.baker_id = self.baker_id
        baker_review_dbmodel.author_id = self.author_id
        baker_review_dbmodel.rating = self.rating
        baker_review_dbmodel.description = self.description
        return baker_review_dbmodel

    def fromBakerReviewDbModel(self, baker_review_dbmodel):
        self.review_id = baker_review_dbmodel.id
        self.baker_id = baker_review_dbmodel.baker_id
        self.author_id = baker_review_dbmodel.author_id
        self.rating = baker_review_dbmodel.rating
        self.description = baker_review_dbmodel.description