# swagger_server/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: Bakers hub OpenAPI spec
  version: 0.0.1
servers:
- url: http://localhost:8080/v1
  description: Dev server
paths:
  /bakers:
    get:
      tags:
      - Bakers
      summary: Returns list of bakers
      operationId: get_bakers
      parameters:
      - name: status
        in: query
        description: Bakers' status
        required: false
        style: form
        explode: true
        schema:
          type: string
          enum: [active, not active, banned]
        example: active
      - name: rating
        in: query
        description: Baker's rating
        required: false
        style: form
        explode: true
        schema:
          type: number
        example: 7.5
      responses:
        "200":
          description: Successful response with list of bakers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Baker'
                x-content-type: application/json
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultError'
      x-openapi-router-controller: swagger_server.controllers.bakers_controller
    post:
      tags:
      - Bakers
      summary: Creates a new baker
      operationId: create_baker
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Baker'
        required: true
      responses:
        "200":
          description: Successful response with new baker
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Baker'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultError'
      x-openapi-router-controller: swagger_server.controllers.bakers_controller
  /bakers/{baker_id}:
    get:
      tags:
      - Bakers
      summary: Returns backer with requested ID
      operationId: get_baker_by_id
      parameters:
      - name: baker_id
        in: path
        description: Baker's id
        required: true
        style: simple
        explode: false
        schema:
          type: string
          example: "1"
      responses:
        "200":
          description: Successful response with relevant baker
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Baker'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultError'
      x-openapi-router-controller: swagger_server.controllers.bakers_controller
    put:
      tags:
      - Bakers
      summary: Updates existing baker
      operationId: update_baker_by_id
      parameters:
      - name: baker_id
        in: path
        description: Baker's ID
        required: true
        style: simple
        explode: false
        schema:
          type: string
        example: "1"
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Baker'
        required: true
      responses:
        "200":
          description: Successful response with updated baker
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Baker'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultError'
      x-openapi-router-controller: swagger_server.controllers.bakers_controller
    delete:
      tags:
      - Bakers
      description: Deletes baker with requested ID
      operationId: delete_baker_by_id
      parameters:
      - name: baker_id
        in: path
        description: Baker's ID
        required: true
        style: simple
        explode: false
        schema:
          type: string
        example: "1"
      responses:
        "200":
          description: Successfully deleted
          content:
            application/json: {}
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultError'
      x-openapi-router-controller: swagger_server.controllers.bakers_controller
  /bakers/{baker_id}/reviews:
    get:
      tags:
      - Bakers
      summary: Returns baker reviews
      operationId: get_baker_reviews_by_baker_id
      parameters:
      - name: baker_id
        in: path
        description: Baker's ID
        required: true
        style: simple
        explode: false
        schema:
          type: string
        example: "1"
      responses:
        "200":
          description: Successful response with list of baker reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BakerReview'
                x-content-type: application/json
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultError'
      x-openapi-router-controller: swagger_server.controllers.bakers_controller
    post:
      tags:
      - Bakers
      summary: Creates baker review
      operationId: create_baker_review
      parameters:
      - name: baker_id
        in: path
        description: Baker's ID
        required: true
        style: simple
        explode: false
        schema:
          type: string
        example: "1"
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BakerReview'
        required: true
      responses:
        "200":
          description: Successful response with  baker review
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BakerReview'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultError'
      x-openapi-router-controller: swagger_server.controllers.bakers_controller
    delete:
      tags:
      - Bakers
      summary: Deletes baker review
      operationId: delete_baker_review_by_id
      parameters:
      - name: baker_id
        in: path
        description: Baker's ID
        required: true
        style: simple
        explode: false
        schema:
          type: string
        example: "1"
      - name: review_id
        in: query
        description: Review's ID
        required: true
        style: form
        explode: true
        schema:
          type: string
          example: "1"
      responses:
        "200":
          description: Successfully deleted
          content:
            application/json: {}
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultError'
      x-openapi-router-controller: swagger_server.controllers.bakers_controller
...
components:
  schemas:
    Baker:
      required:
      - contact
      - name
      - status
      type: object
      properties:
        baker_id:
          type: string
          example: "1"
        name:
          type: string
          example: Mila
        contact:
          type: string
          example: "79200123456"
        status:
          type: string
          enum:
          - active
          - not active
          - deleted
        rating:
          type: number
          example: 0
        finished_orders:
          type: integer
          example: 0
      example:
        finished_orders: 0
        contact: "79200123456"
        name: Mila
        rating: 0
        baker_id: "1"
        status: active
...
    BakerReview:
      required:
      - author_id
      - description
      - rating
      type: object
      properties:
        review_id:
          type: string
          example: "1"
        baker_id:
          type: string
        author_id:
          type: string
          example: "1"
        rating:
          type: integer
          example: 9
          enum:
          - 1
          - 2
          - 3
          - 4
          - 5
          - 6
          - 7
          - 8
          - 9
          - 10
        description:
          type: string
          example: very good very well
      example:
        review_id: "1"
        rating: 9
        description: very good very well
        baker_id: baker_id
        author_id: "1"
...
    DefaultError:
      required:
      - code
      - message
      type: object
      properties:
        code:
          type: integer
        message:
          type: string